Move Figure class and helper funcs to figure.py
